+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Convenience properties on the return value of broadcasted calls
 - `.broadcast{.once,.onceBy,}.orThrow()`: aggregate method that throws if any node failed and returns the array of results otherwise.
 - `nodes` property in the third ("tools") argument of procedure implementations, containing the set of all available node IDs.
+- `Client#destroy` to terminate workers and disconnect event listeners.
 
 ## [0.18.0] - 2026-01-08